Apply part of review suggestions
