+++ crates/sncast/src/starknet_commands/multicall/contract_registry.rs
@@ -8,8 +8,11 @@ use starknet_types_core::felt::Felt;
 use std::collections::{HashMap, hash_map::Entry};
 
 /// Registry for managing contract information during multicall execution.
-/// It maintains mappings between user-defined ids, contract addresses, class hashes, and contract classes
-/// to facilitate reference and reuse of deployed contracts across multiple calls in a multicall sequence.
+/// It stores the following mappings:
+/// - user-defined ids to contract addresses
+/// - contract addresses to class hashes
+/// - class hashes to contract classes
+/// to allow referencing and re-using deployed contracts across multiple calls in a multicall sequence.
 pub struct ContractRegistry {
     id_to_address: HashMap<String, Felt>,
     address_to_class_hash: HashMap<Felt, Felt>,
@@ -45,8 +48,8 @@ impl ContractRegistry {
     /// Retrieves the class hash associated with the given contract address.
     /// Checks the local cache first, and fetches from the provider if not cached.
     pub async fn get_class_hash_by_address(&mut self, address: &Felt) -> Result<Felt> {
-        if let Some(hash) = self.address_to_class_hash.get(address) {
-            return Ok(*hash);
+        if let Some(hash) = self.get_class_hash_by_address_local(address) {
+            return Ok(hash);
         }
         let class_hash = get_class_hash_by_address(&self.provider, *address)
             .await
