--- crates/sncast/src/starknet_commands/utils/dry_run.rs
@@ -1,1 +1,0 @@
-pub fn run_with_dry_run() -> () {}
